Move CiteDialog to components folder
