+++ asreview/webapp/src/Components/index.js
@@ -2,6 +2,7 @@ export { default as ActionsFeedbackBar } from "./ActionsFeedbackBar";
 export { default as AppBarWithinDialog } from "./AppBarWithinDialog";
 export { default as BoxErrorHandler } from "./BoxErrorHandler";
 export { default as CardErrorHandler } from "./CardErrorHandler";
+export { default as CiteDialog } from "./CiteDialog";
 export { default as ConfirmAccount } from "./ConfirmAccount";
 export { default as DialogErrorHandler } from "./DialogErrorHandler";
 export { default as DrawerItem } from "./DrawerItem";
